Refactored settings-related classes. Moved non-public API classes to 'internal.settings' package.
