+++ instancio-tests/instancio-core-tests/src/test/resources/instancio.properties
@@ -39,4 +39,4 @@ string.nullable=false
 subtype.java.util.Collection=java.util.ArrayList
 subtype.java.util.List=java.util.ArrayList
 subtype.java.util.Map=java.util.HashMap
-subtype.java.util.SortedMap=java.util.TreeMap
+subtype.java.util.SortedMap=java.util.concurrent.ConcurrentSkipListMap
